refactor: Use standard lazy loading for repl runner & editor
