+++ src/components/content-region/index.jsx
@@ -5,8 +5,8 @@ import style from './style.module.css';
 import { useTranslation } from '../../lib/i18n';
 import { TocContext } from '../table-of-contents';
 import { prefetchContent } from '../../lib/use-content';
-import { preloadRepl } from '../../lib/use-repl';
-import { Repl, TutorialPage } from '../routes';
+import { preloadRepl } from '../../lib/repl';
+import { ReplPage, TutorialPage } from '../routes';
 
 const COMPONENTS = {
 	...widgets,
@@ -16,7 +16,7 @@ const COMPONENTS = {
 
 			props.onMouseOver = () => {
 				if (props.href.startsWith('/repl?code')) {
-					Repl.preload();
+					ReplPage.preload();
 					preloadRepl();
 				} else if (props.href.startsWith('/tutorial')) {
 					TutorialPage.preload();
