+++ src/components/routes.jsx
@@ -5,7 +5,7 @@ import { DocPage } from './controllers/doc-page';
 import { NotFound } from './controllers/not-found';
 import { navRoutes } from '../lib/route-utils';
 
-export const Repl = lazy(() => import('./controllers/repl-page'));
+export const ReplPage = lazy(() => import('./controllers/repl-page'));
 export const BlogPage = lazy(() => import('./controllers/blog-page'));
 export const TutorialPage = lazy(() => import('./controllers/tutorial-page'));
 
@@ -26,7 +26,7 @@ export default function Routes() {
 					.filter(route => !route.startsWith('/guide'))
 					.filter(route => !route.startsWith('/tutorial'))
 					.map(route => {
-						const component = route === '/repl' ? Repl : Page;
+						const component = route === '/repl' ? ReplPage : Page;
 						return <Route key={route} path={route} component={component} />;
 					})}
 				<Route path="/tutorial/:step?" component={TutorialPage} />
